composetest/requirements.txt
flask
redis
pymysql
mysql-connector
mysql-connector-python
